+++ EXILED/Exiled.Events/EventArgs/Player/Scp1576TransmissionEndedEventArgs.cs
@@ -1,11 +1,11 @@
 ﻿// -----------------------------------------------------------------------
-// <copyright file="TransmissionEndedEventArgs.cs" company="ExMod Team">
+// <copyright file="Scp1576TransmissionEndedEventArgs.cs" company="ExMod Team">
 // Copyright (c) ExMod Team. All rights reserved.
 // Licensed under the CC BY-SA 3.0 license.
 // </copyright>
 // -----------------------------------------------------------------------
 
-namespace Exiled.Events.EventArgs.Scp1576
+namespace Exiled.Events.EventArgs.Player
 {
     using API.Features;
     using API.Features.Items;
@@ -15,14 +15,14 @@ namespace Exiled.Events.EventArgs.Scp1576
     /// <summary>
     /// Contains all information after a SCP-1576 transmission has ended.
     /// </summary>
-    public class TransmissionEndedEventArgs : IPlayerEvent
+    public class Scp1576TransmissionEndedEventArgs : IPlayerEvent
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="TransmissionEndedEventArgs"/> class.
+        /// Initializes a new instance of the <see cref="Scp1576TransmissionEndedEventArgs"/> class.
         /// </summary>
         /// <param name="player"><inheritdoc cref="Player"/></param>
         /// <param name="scp1576Item"><inheritdoc cref="Scp1576"/></param>
-        public TransmissionEndedEventArgs(Player player, Scp1576Item scp1576Item)
+        public Scp1576TransmissionEndedEventArgs(Player player, Scp1576Item scp1576Item)
         {
             Player = player;
             Scp1576 = Item.Get<Scp1576>(scp1576Item);
