+++ EXILED/Exiled.Events/Patches/Events/Player/Scp1576TransmissionEnded.cs
@@ -1,30 +1,29 @@
 ﻿// -----------------------------------------------------------------------
-// <copyright file="TransmissionEnded.cs" company="ExMod Team">
+// <copyright file="Scp1576TransmissionEnded.cs" company="ExMod Team">
 // Copyright (c) ExMod Team. All rights reserved.
 // Licensed under the CC BY-SA 3.0 license.
 // </copyright>
 // -----------------------------------------------------------------------
 
-namespace Exiled.Events.Patches.Events.Scp1576
+namespace Exiled.Events.Patches.Events.Player
 {
     using System.Collections.Generic;
     using System.Reflection.Emit;
 
-    using Exiled.API.Features;
     using Exiled.API.Features.Pools;
     using Exiled.Events.Attributes;
-    using Exiled.Events.EventArgs.Scp1576;
+    using Exiled.Events.EventArgs.Player;
     using HarmonyLib;
     using InventorySystem.Items.Usables.Scp1576;
 
     using static HarmonyLib.AccessTools;
 
     /// <summary>
-    /// Patches <see cref="Scp1576Item.ServerStopTransmitting"/> to add <see cref="Handlers.Scp1576.TransmissionEnded"/> event.
+    /// Patches <see cref="Scp1576Item.ServerStopTransmitting"/> to add <see cref="Handlers.Player.Scp1576TransmissionEnded"/> event.
     /// </summary>
-    [EventPatch(typeof(Handlers.Scp1576), nameof(Handlers.Scp1576.TransmissionEnded))]
+    [EventPatch(typeof(Handlers.Player), nameof(Handlers.Player.Scp1576TransmissionEnded))]
     [HarmonyPatch(typeof(InventorySystem.Items.Usables.Scp1576.Scp1576Item), nameof(Scp1576Item.ServerStopTransmitting))]
-    public class TransmissionEnded
+    public class Scp1576TransmissionEnded
     {
         private static IEnumerable<CodeInstruction> Transpiler(IEnumerable<CodeInstruction> instructions, ILGenerator generator)
         {
@@ -39,16 +38,16 @@ namespace Exiled.Events.Patches.Events.Scp1576
                         // Player.Get(base.Owner)
                         new CodeInstruction(OpCodes.Ldarg_0),
                         new CodeInstruction(OpCodes.Callvirt, PropertyGetter(typeof(Scp1576Item), nameof(Scp1576Item.Owner))),
-                        new CodeInstruction(OpCodes.Call, Method(typeof(API.Features.Player), nameof(Player.Get), new[] { typeof(ReferenceHub) })),
+                        new CodeInstruction(OpCodes.Call, Method(typeof(API.Features.Player), nameof(API.Features.Player.Get), new[] { typeof(ReferenceHub) })),
 
                         // this
                         new CodeInstruction(OpCodes.Ldarg_0),
 
                         // ServerStopTransmittingEventArgs ev = new(Player, this)
-                        new CodeInstruction(OpCodes.Newobj, GetDeclaredConstructors(typeof(TransmissionEndedEventArgs))[0]),
+                        new CodeInstruction(OpCodes.Newobj, GetDeclaredConstructors(typeof(Scp1576TransmissionEndedEventArgs))[0]),
 
                         // OnTransmisionEnded(ev)
-                        new CodeInstruction(OpCodes.Call, Method(typeof(Handlers.Scp1576), nameof(Handlers.Scp1576.OnTransmisionEnded))),
+                        new CodeInstruction(OpCodes.Call, Method(typeof(Handlers.Player), nameof(Handlers.Player.OnScp1576TransmissionEnded))),
                     });
 
                 for (int i = 0; i < newInstructions.Count; i++)
